[#78] patch_task_definition를 서비스 계층으로 재작성
