+++ src/actions/mod.rs
@@ -4,7 +4,6 @@ pub mod delete_task_definition;
 pub mod list_schedules;
 pub mod list_task_definition;
 pub mod patch_schedule;
-pub mod patch_task_definition;
 pub mod run_pending_job;
 pub mod stop_job;
 pub mod submit_job;
