+++ src/domain/task_definition/service.rs
@@ -2,8 +2,8 @@ use std::sync::Arc;
 
 use super::{
     TaskDefinitionRepository,
-    dao::{CreateTaskDefinitionParams, ListTaskDefinitionsParams},
-    dto::CreateDefinitionRequest,
+    dao::{CreateTaskDefinitionParams, ListTaskDefinitionsParams, PatchTaskDefinitionParams},
+    dto::{CreateDefinitionRequest, PatchDefinitionRequest},
     entities,
 };
 
@@ -65,5 +65,38 @@ impl super::TaskDefinitionService for TaskDefinitionServiceImpl {
             .await?;
 
         Ok(task_definition_id)
+    }
+
+    async fn patch_task_definition(&self, request: PatchDefinitionRequest) -> anyhow::Result<()> {
+        // version이 없다면 동일한 이름의 task definition이 있는지 확인
+
+        let task_definitions = self
+            .task_definition_repository
+            .list_task_definitions(ListTaskDefinitionsParams {
+                task_definition_ids: vec![request.task_definition_id],
+                limit: Some(1),
+                ..Default::default()
+            })
+            .await?;
+
+        if task_definitions.is_empty() {
+            return Err(anyhow::anyhow!("Task definition not found"));
+        }
+
+        let _ = self
+            .task_definition_repository
+            .patch_task_definition(PatchTaskDefinitionParams {
+                task_definition_id: request.task_definition_id,
+                image: request.request.image,
+                command: request.request.command,
+                args: request.request.args,
+                env: request.request.env,
+                memory_limit: request.request.memory_limit,
+                cpu_limit: request.request.cpu_limit,
+                ..Default::default()
+            })
+            .await;
+
+        Ok(())
     }
 }
